+++ packages/illustrations/src/__generated__/pictogram/data/names.ts
@@ -215,6 +215,7 @@ const names: PictogramName[] = [
   'coinbaseOneShield',
   'coinbaseOneTrade',
   'coinbaseOneTrusted',
+  'coinbaseOneUnlimtedRewards',
   'coinbaseWalletApp',
   'coldStorageCheck',
   'collectionOfAssets',