+++ packages/illustrations/src/__generated__/pictogram/types/PictogramName.ts
@@ -207,6 +207,7 @@ export type PictogramName =
   | 'coinbaseOneShield'
   | 'coinbaseOneTrade'
   | 'coinbaseOneTrusted'
+  | 'coinbaseOneUnlimtedRewards'
   | 'coinbaseWalletApp'
   | 'coinFocus'
   | 'coinShare'